feat: enhance order processing with user validation, cart updates, and improved error handling

+++ src/main/java/com/project2/BookStore/dto/OrderResponseDTO.java
@@ -35,7 +35,7 @@ public class OrderResponseDTO {
         private String bookTitle;
         private String bookImage;
         private Integer quantity;
-        private Double price;
-        private Double subtotal;
+        private Long price;
+        private Long subtotal;
     }
 }
